moved import instrumenter methods to dedicated class and rewrote destination imports changer
